Move PNC Update Dataset serialisation to lib directory
