+++ packages/core/lib/serialise/pnc-update-dataset-xml/serialiseToXml.test.ts
@@ -1,8 +1,8 @@
 import fs from "fs"
 import "jest-xml-matcher"
-import { parsePncUpdateDataSetXml } from "../../parse/parsePncUpdateDataSetXml"
-import serialiseToXml from "./serialiseToXml"
+import { parsePncUpdateDataSetXml } from "../../../phase2/parse/parsePncUpdateDataSetXml"
 import type { PncUpdateDataset } from "../../../types/PncUpdateDataset"
+import serialiseToXml from "./serialiseToXml"
 
 describe("serialiseToXml", () => {
   it.each([
